김관호 : 리소스 로드 구조 추가

+++ WinAPI/Client/struct.h
@@ -14,7 +14,7 @@ public:
 
 	float Length()
 	{
-		return sqrt(x * x + y * y);
+		return static_cast<float>(sqrt(x * x + y * y));
 	}
 
 	Vec2& Normalize()
